all: better, more consistent names
Names of various things, especially functions or variations, were inconsistent and confusing, and some types had names that were a bit too short. Document precise definitions of terms and convert everything to use them, and rename F => Func, R => Render, and P => Pt.

Additionally, remove the xirho.Param type. Nothing used it, because fapi.Param is the same thing but with semantics. As a result of the removal of xirho.Param, the old xirho.Func type which was a struct wrapper around the function interface type can be removed.